+++ WooCommerce/WooCommerceTests/ViewRelated/Dashboard/Free Trial/StorePlanBannerPresenterTests.swift
@@ -32,41 +32,3 @@ final class StorePlanBannerPresenterTests: XCTestCase {
         }
     }
 }
-
-private final class MockStorePlanSynchronizer: StorePlanSynchronizing {
-    var planState: AnyPublisher<StorePlanSyncState, Never> {
-        planStateSubject.eraseToAnyPublisher()
-    }
-
-    private let planStateSubject = CurrentValueSubject<StorePlanSyncState, Never>(.notLoaded)
-
-    var site: Site?
-
-    func reloadPlan() {
-        // no-op
-    }
-
-    func setState(_ state: StorePlanSyncState) {
-        planStateSubject.send(state)
-    }
-}
-
-private final class MockConnectivityObserver: ConnectivityObserver {
-    @Published private(set) var currentStatus: ConnectivityStatus = .unknown
-
-    var statusPublisher: AnyPublisher<ConnectivityStatus, Never> {
-        $currentStatus.eraseToAnyPublisher()
-    }
-
-    func startObserving() {
-        // no-op
-    }
-
-    func stopObserving() {
-        // no-op
-    }
-
-    func setStatus(_ status: ConnectivityStatus) {
-        currentStatus = status
-    }
-}
